requirements.txt
flask
blinker
mysql.connector
ddtrace
datadog
